fix circular dependencies between Editor and codecs
