+++ packages/core/src/index.ts
@@ -96,7 +96,9 @@ export { default as CodecRegistry } from './serialization/CodecRegistry';
 export { default as ObjectCodec } from './serialization/ObjectCodec';
 export * from './serialization/ModelXmlSerializer';
 export * from './serialization/codecs';
+export * from './serialization/codecs/_export-model-codecs';
 export * from './serialization/register';
+export * from './serialization/register-model-codecs';
 
 export { default as ActorShape } from './view/geometry/ActorShape';
 export { default as LabelShape } from './view/geometry/node/LabelShape';
